% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Main}
\alias{Main}
\title{Main model fitting function}
\usage{
Main(
  Y,
  X,
  iteration,
  burnin,
  nchain,
  tau = NULL,
  nu0 = 5,
  sigma0 = NULL,
  alpha1p = 2,
  p1o = NULL,
  alpha0p = 2,
  p0o = NULL,
  alpha1 = NULL,
  alpha2 = NULL,
  dependency = TRUE,
  strength = TRUE
)
}
\arguments{
\item{Y}{A list of response values in regressions}

\item{X}{A list of design matrices in regressions}

\item{iteration}{The number of MCMC iterations}

\item{burnin}{The number of burn-in iterations}

\item{nchain}{The number of MCMC chains}

\item{dependency}{Whether to use the ISP prior; A value of TRUE means creating dependencies among SNPs}

\item{strength}{Whether to borrow strength across regressions using MRF prior}
}
\value{
A large list of all the parameter values
}
\description{
This is the main model fitting function for Bayesian framework with ISP and MRF priors.
}
\details{
This function is used to fit a Baeysian regression model with ISP and MRF priors.
}
\examples{
library(mvtnorm)
K = 3
P = 500
Nk = c(50, 100, 150)
GAMMA = matrix(0, nrow = K, ncol = P)
GAMMA[1, 1:8] = 1
GAMMA[2, 1:8] = 1
GAMMA[3, 1:8] = 1
Y = lapply(Nk, function(l) rep(0, l))
X = lapply(Nk, function(l) matrix(0, l, P))
sigma = lapply(Nk, function(l) rep(0, l))
SigmaX = 0.05 ^ abs(matrix(1:P - 1, nrow = P, ncol = P, byrow = TRUE) - (1:P - 1))
BETA = matrix(0, nrow = K, ncol = P)
sigma = lapply(Nk, function(l) rep(0, l))

for (k in 1:K) {
 X[[k]] = matrix(rmvnorm(Nk[k], sigma = SigmaX), nrow = Nk[k], byrow = TRUE)
 BETA[k, which(GAMMA[k, ] == 1)] = c(2, -2, 2, -2, 2, -2, 2, -2)
 sigma[[k]] = rnorm(Nk[k], 0, 1)
 Y[[k]] = X[[k]] \%*\% (GAMMA[k, ] * BETA[k, ]) + sigma[[k]]
 Y[[k]] = as.numeric(Y[[k]] > 0)
}

t0 = Sys.time()
res = Main(Y, X, iteration = 5000, burnin = 3500, nchain = 2)
Sys.time() - t0

}
