% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{Predict}
\alias{Predict}
\title{Cross-validated prediction function}
\usage{
Predict(res, Y, X, burnin, nfold = 5, ich = 1)
}
\arguments{
\item{res}{The model output from the Main() function}

\item{Y}{A list of response values in regressions}

\item{X}{A list of design matrices in regressions}

\item{burnin}{The number of burn-in iterations}

\item{nfold}{The number of folds for cross validation}

\item{ich}{The index of chain used for prediction (e.g., ich = 1)}
}
\value{
A list of predicted probabilities for each subject to be a case
}
\description{
This is the prediction function with built-in cross validation.
}
\details{
This function performs cross-validation and generates predicted values.
}
\examples{
Predict(res, Y, X, 3500)

}
